Tracer refactoring and lots of improvements
