+++ modules/tracer/backend/src/main/scala/TracerServer.scala
@@ -29,7 +29,6 @@ import fs2.text
 import jsonrpclib.Codec
 import jsonrpclib.Payload
 import jsonrpclib.fs2.lsp
-import langoustine.tracer.{Message as LspMessage}
 import org.http4s.*
 import org.http4s.HttpApp
 import org.http4s.dsl.*
@@ -64,9 +63,10 @@ class TracerServer private (
       val run = Server(
         wbs =>
           ErrorHandling
-            .httpApp(handleErrors(app(wbs, state, summary))),
+            .httpApp(handleErrors(api(wbs, state, summary))),
         config
       )
+
       val dump = dumpRequests(state)
         .concurrently(dumpResponses(state))
         .concurrently(dumpLogs(state))
@@ -77,8 +77,10 @@ class TracerServer private (
       dump *> run
     }
 
-  private def dumpRequests(state: State)  = dump(in, state, Direction.ToServer)
-  private def dumpResponses(state: State) = dump(out, state, Direction.ToClient)
+  private def dumpRequests(state: State) =
+    state.hydrateFrom(in, Direction.ToServer)
+  private def dumpResponses(state: State) =
+    state.hydrateFrom(out, Direction.ToClient)
 
   private def dumpLogs(state: State) =
     err
@@ -100,131 +102,6 @@ class TracerServer private (
 
   import org.http4s.dsl.io.*
 
-  private def app(
-      wbs: WebSocketBuilder2[IO],
-      state: State,
-      summary: Summary
-  ) =
-    import state.*
-    val apiRoutes = HttpRoutes.of[IO] {
-      case GET -> Root / "summary" =>
-        Ok(summary)
-
-      case GET -> Root / "ws" / "events" =>
-        val continuous =
-          messages.discrete
-            .as(TracerEvent.Update)
-            .merge(messages.discrete.as(TracerEvent.Update))
-            .merge(ch.stream.map(l => TracerEvent.LogLine(l)))
-
-        val out = fs2.Stream.eval(logBuf.get).map(TracerEvent.LogLines(_))
-
-        val updates =
-          (out ++ continuous)
-            .map(ev =>
-              WebSocketFrame.Text(writeToStringReentrant[TracerEvent](ev))
-            )
-
-        val pings =
-          fs2.Stream.repeatEval(IO(WebSocketFrame.Ping())).metered(1.second)
-
-        wbs
-          .withFilterPingPongs(false)
-          .build(updates.merge(pings), _.map(_ => ()))
-
-      case GET -> Root / "all" =>
-        messages.get.map(_.sortBy(_.timestamp).map(_.decoded)).flatMap(Ok(_))
-
-      case GET -> Root / "logs" =>
-        logBuf.get.flatMap(Ok(_))
-
-      case GET -> Root / "raw" / "all" =>
-        messages.get.map(_.sortBy(_.timestamp).map(_.raw)).flatMap(Ok(_))
-
-      case GET -> Root / "snapshot" :? SnapshotNameMatcher(nameMaybe) =>
-        given JsonValueCodec[Vector[ReceivedMessage]] = JsonCodecMaker.make
-
-        val filename = nameMaybe
-          .getOrElse("langoustine-tracer-snapshot")
-          + ".jsonl.gz"
-
-        messages.get.map(_.sortBy(_.timestamp)).flatMap { received =>
-          val bytes = fs2.Stream
-            .emits(received)
-            .covary[IO]
-            .evalMap(m => IO(writeToStringReentrant(m)))
-            .intersperse("\n")
-            .through(fs2.text.utf8.encode)
-            .through(fs2.compression.Compression[IO].gzip())
-
-          Ok(
-            bytes,
-            org.http4s.headers.`Content-Disposition`(
-              s"attachment",
-              Map(CIString("filename") -> filename)
-            )
-          )
-        }
-
-      case GET -> Root / "raw" / "request" / id =>
-        val asLong = id.toLongOption.map(MessageId.NumberId.apply)
-        val asStr  = MessageId.StringId(id)
-
-        inline def idMatches(idOpt: Option[MessageId]) =
-          (asLong == idOpt || idOpt.contains(asStr))
-
-        messages.get.flatMap { messages =>
-          val found = messages.collectFirst {
-            case rm if idMatches(rm.raw.id) && rm.raw.method.nonEmpty =>
-              rm.raw
-          }
-
-          found match
-            case Some(rm) => Ok(rm)
-            case None     => NotFound()
-        }
-
-      case GET -> Root / "raw" / "response" / id =>
-        val asLong = id.toLongOption.map(MessageId.NumberId.apply)
-        val asStr  = MessageId.StringId(id)
-
-        inline def idMatches(idOpt: Option[MessageId]) =
-          (asLong == idOpt || idOpt.contains(asStr))
-
-        messages.get.flatMap { messages =>
-          val found = messages.collectFirst {
-            case rm if idMatches(rm.raw.id) && rm.raw.method.isEmpty =>
-              rm.raw
-          }
-
-          found match
-            case Some(rm) => Ok(rm)
-            case None     => NotFound()
-        }
-
-      case GET -> Root / "raw" / "notification" / id =>
-        val asStr = MessageId.StringId(id)
-
-        inline def idMatches(idOpt: Option[MessageId]) =
-          idOpt.contains(asStr)
-
-        messages.get.flatMap { messages =>
-          val found = messages.collectFirst {
-            case rm if idMatches(rm.raw.id) => rm.raw
-          }
-
-          found match
-            case Some(rm) => Ok(rm)
-            case None     => NotFound()
-        }
-    }
-
-    Router("/api" -> apiRoutes) <+> Static.routes
-  end app
-
-  object SnapshotNameMatcher
-      extends OptionalQueryParamDecoderMatcher[String]("name")
-
   private def Server(
       app: WebSocketBuilder2[IO] => HttpApp[IO],
       config: BindConfig
@@ -240,21 +117,6 @@ class TracerServer private (
 end TracerServer
 
 object TracerServer:
-  given msgCodec: JsonValueCodec[Vector[LspMessage]] = JsonCodecMaker.make
-  given rawCodec: JsonValueCodec[Vector[RawMessage]] = JsonCodecMaker.make
-
-  given JsonValueCodec[Vector[String]] = JsonCodecMaker.make
-  given jsonEncoder[T: JsonValueCodec]: EntityEncoder[IO, T] =
-    EntityEncoder
-      .byteArrayEncoder[IO]
-      .contramap((data: T) => writeToArray(data))
-      .withContentType(
-        `Content-Type`(MediaType.application.json, Some(Charset.`UTF-8`))
-      )
-
-  given jsonDecoder[T: JsonValueCodec]: EntityDecoder[IO, T] =
-    EntityDecoder.byteArrayDecoder[IO].map(readFromArray[T](_))
-
   def create(
       in: fs2.Stream[IO, Byte],
       out: fs2.Stream[IO, Byte],
@@ -262,99 +124,10 @@ object TracerServer:
   ): TracerServer =
     TracerServer(in, out, err)
 
-  case class State(
-      ch: Channel[IO, String],
-      messages: SignallingRef[IO, Vector[ReceivedMessage]],
-      logBuf: Ref[IO, Vector[String]],
-      responseIdMapping: Ref[IO, Map[MessageId, String]],
-      random: UUIDGen[IO]
-  )
-
-  object State:
-    def create =
-      for
-        messages          <- SignallingRef[IO].of(Vector.empty[ReceivedMessage])
-        logBuf            <- IO.ref(Vector.empty[String])
-        ch                <- Channel.bounded[IO, String](128)
-        responseIdMapping <- IO.ref(Map.empty[MessageId, String])
-      yield State(ch, messages, logBuf, responseIdMapping, UUIDGen[IO])
-  end State
-
-  def dump(stream: fs2.Stream[IO, Byte], state: State, direction: Direction) =
-    import state.*
-    stream
-      .through(lsp.decodePayloads[IO])
-      .evalMap(decodeRaw(_, state))
-      .collect { case Some(v) =>
-        v
-      }
-      .evalTap { rawMessage =>
-        random.randomUUID
-          .map(u => MessageId.StringId("generated-" + u.toString))
-          .flatMap { generatedId =>
-            val receivedRawMessage = rawMessage.copy(value =
-              rawMessage.value
-                .copy(id = rawMessage.value.id.orElse(Some(generatedId)))
-            )
-
-            val msg = LspMessage
-              .from(rawMessage.value, direction, generatedId)
-            val receivedMsg = msg.map(Received(rawMessage.timestamp, _))
-
-            LspMessage
-              .from(rawMessage.value, direction, generatedId)
-              .map { lspMessage =>
-
-                val withUpdatedMapping = lspMessage match
-                  case LspMessage.Request(method, id, _) =>
-                    responseIdMapping
-                      .update(_.updated(id, method))
-                      .as(lspMessage)
-
-                  case LspMessage.Response(id, _) =>
-                    messages.update { received =>
-                      received.collect {
-                        case ReceivedMessage(ts, raw, req: LspMessage.Request)
-                            if req.id == id =>
-                          ReceivedMessage(ts, raw, req.copy(responded = true))
-                        case other => other
-                      }
-                    } *>
-                      responseIdMapping.get
-                        .map(_.get(id))
-                        .map(LspMessage.Response(id, _))
-
-                  case other => IO.pure(other)
-
-                withUpdatedMapping
-                  .flatMap { msg =>
-                    messages.update(
-                      _ :+ ReceivedMessage(
-                        rawMessage.timestamp,
-                        rawMessage.value,
-                        msg
-                      )
-                    )
-                  }
-
-              }
-              .getOrElse(IO.unit)
-
-          }
-      }
-  end dump
-
-  case class Received[T](timestamp: Long, value: T)
-  object Received:
-    def capture[T](t: T) = IO.monotonic.map(lng => Received(lng.toMillis, t))
-    def captureF[T](t: IO[T]) =
-      IO.monotonic.flatMap(lng => t.map(Received(lng.toMillis, _)))
-
-  def decodeRaw(p: Payload, st: State) =
-    IO(Codec.decode[RawMessage](Some(p))).flatMap {
-      case Left(err) =>
-        st.ch.send(s"[Tracer] hard failed to decode $p, error: $err").as(None)
-      case Right(v) => Received.capture(v).map(Option.apply)
-    }
-
 end TracerServer
+
+case class Received[T](timestamp: Long, value: T)
+object Received:
+  def capture[T](t: T) = IO.monotonic.map(lng => Received(lng.toMillis, t))
+  def captureF[T](t: IO[T]) =
+    IO.monotonic.flatMap(lng => t.map(Received(lng.toMillis, _)))
