+++ modules/tracer/backend/src/main/scala/routes.static.scala
@@ -33,7 +33,9 @@ object Static:
 
     HttpRoutes.of[IO] {
       case req @ GET -> Root / "assets" / filename
-          if filename.endsWith(".js") || filename.endsWith(".js.map") =>
+          if filename.endsWith(".js") ||
+            filename.endsWith(".js.map") ||
+            filename.endsWith(".svg") =>
         StaticFile
           .fromResource[IO](
             Paths.get("assets", filename).toString,
