+++ modules/tracer/frontend/src/main/scala/page.commands.scala
@@ -24,14 +24,15 @@ import com.raquo.airstream.state.Var
 def commandTracer(
     bus: EventBus[Double],
     commandFilter: Var[Option[String]],
-    messagesState: Var[Vector[Message]],
-    showing: Var[Option[Message]],
-    mode: Var[JsonMode]
+    messagesState: Var[Vector[LspMessage]],
+    showing: Var[Option[LspMessage]],
+    mode: Var[JsonMode],
+    modalBus: EventBus[ModalCommand]
 ) =
-  import Message.*
+  import LspMessage.*
   div(
     Styles.commandTracer.container,
-    jsonViewer(showing, mode),
+    jsonViewer(showing, mode, modalBus),
     div(
       width <-- showing.signal
         .map(_.isDefined)
